+++ tests/testthat/test-naive_model.R
@@ -13,3 +13,40 @@ familiar:::integrated_test(
   parallel = FALSE,
   debug = debug_flag
 )
+
+
+# Test for additional errors in situations where a naive model is created due
+# to lack of actual information among the features.
+data <- familiar:::test_create_random_data("continuous", seed = 19L)
+
+# Though the dataset should generally lead to a naive model being formed, this
+# is not guaranteed. We iterate until it works.
+results <- NULL
+while (!is(results$familiarModel, "familiarNaiveModel")) {
+  results <- familiar::summon_familiar(
+    data = data,
+    vimp_method = "mim",
+    learner = "glm_gaussian",
+    hyperparameter = list("glm_gaussian" = list("sign_size" = c(4, 5, 6))),
+    outcome_type = "continuous",
+    experimental_design = "fs+mb",
+    estimation_type = "point",
+    verbose = debug_flag
+  )
+}
+
+testthat::test_that(
+  "naive models are correctly specified", 
+  {
+    testthat::expect_s4_class(results$familiarModel, "familiarNaiveModel")
+    testthat::expect_length(results$familiarModel@required_features, 0L)
+    testthat::expect_length(results$familiarModel@model_features, 0L)
+    testthat::expect_length(results$familiarModel@novelty_features, 0L)
+    
+    # Internally, naive models are forced to resolve to not train novelty
+    # detectors.
+    testthat::expect_s4_class(results$familiarModel@novelty_detector, "familiarNoneNoveltyDetector")
+    testthat::expect_length(results$familiarModel@novelty_detector@model_features, 0L)
+    testthat::expect_length(results$familiarModel@novelty_detector@required_features, 0L)
+  }
+)