+++ NEWS.md
@@ -1,3 +1,15 @@
+# Version 2.0.2 (Cackling Caribou)
+
+## Fixes
+
+- Fixed an (external) error that was caused by attempting to predict values using
+  `random_forest_rfsrc` or `svm` for prospective data without known outcome.
+
+- Fixed an issue that could occur due to training novelty detectors for naive 
+  models. This could cause a mismatch availability of features in the dataset 
+  for predicting novelty. For naive models, novelty predictors are now no 
+  longer trained.
+
 # Version 2.0.1 (Brilliant Bat)
 
 ## Minor changes